Change componentParser into a class
