+++ packages/core/test/unit/ComponentParser.data.js
